Completed User Story BL055 - normal and reference range calculations. The coefficients for these equations are currently unprotected though, so I modified the Doxyfile-Interface-Headers-Stubs to only include documentation and not header files. This will protect the coefficents from TM until they can be made into native code.
git-svn-id: https://sfs.atcorau.local/svn/SoftwareRepository/trunk@214 a8a9b14d-6b92-5b4b-9153-4cd90bb22c97

+++ PN022BLL/biz/TestBiz/BizCircularBufferTest.cpp
@@ -204,10 +204,10 @@ namespace TestBiz {
 public: [TestMethod]
 		void ReadBufferTest()
 		{
-			unsigned int bufferSize = 10; 
+			unsigned short bufferSize = 10; 
 			BizCircularBuffer^  target = (gcnew BizCircularBuffer(bufferSize)); 
 			cli::array< unsigned short >^  buffer; 
-			unsigned int bufferSizeExpected = 10; 
+			unsigned short bufferSizeExpected = 10; 
 			bufferSize = 0;
 			unsigned short startIndexExpected = 0; 
 			unsigned short startIndex = 0;
